Add ClientStore and store generated SystemInfo.
Add the ClientStore and add support for storing SystemInfo JSON blobs
in a hierarchy so that they can be later processed to simulate client
registrations.

Update Client to add a Hostname() call that combines the client type
and client id to generate a hostname.

Update build rules to store built binaries in a configurable location
defaulting to the out directory under the repo.

Update Dockerfile to build the tools in this repo and SUSE/connect-ng
and store them under /app/bin.