Add Makefile rules for testing purposes.
The client-register rules will trigger building of the testing image,
as well as generation of candidate client hwinfo data, and will then
use the image to register those clients with the target RMT.

The client-tester run can be used to start the testing image as an
interactive session.

Updated the entrypoint script message to indicate where to find the
resources copied or mounted into the image.