Add helper scripts for querying RMT DB
Add the bin/rmt-db-query helper script to simplify querying the RMT
DB associated with a local SUSE/rmt developer deployment.

Add additional helper scripts that leverage the rmt-db-query script
to determine the size of the systems table, and the total size of
the system_information column content within the systems table.