Add rmt-cli helper
The bin/rmt-cli helper allows for easier interaction and management
of the locally running developer deployment RMT.

Update client-tester rule to report the table and system information
sizes after registering clients.

Tweaked the bin/rmt-db-query to allow for passing arguments lines
with flags as an alternative to queries.

Update README.md to reflect addition of rmt-cli.