Misc Enhancments
Modified the rmt-hwinfo-clientctl helper to send the profile data
entries in a new data_profiles entry in the ExtraData blob, which
means that they will be added to the main params block of requests
sent to the RMT, rather than in the hwinfo blob. This should avoid
negative sideeffects of using a newer suseconnect with an older RMT,
as older RMTs will ignore these extra params.

Re-work the client store to create a full hierarchy of directories
with the leaf directory containing multiple files as needed, such
as the sysinfo.json. Added some initial plumbing to support loading
registration info from a reginfo.json file.

Also added ReadFile and WriteFile routines to the clientstore to
simplify access to files in the client data stores.

Add a HwInfoStats.json file to the top-level directory of generated
client data stores to includes statistics about the generated client
entries, such as the number of unique data profiles generated, and
the amount of data storage space that would be saved by aggregating
them in the proposed way. Also includes an estimate of the net space
savings in the RMT DB by using the new scheme.

Increased the variability in the generated client profiles in many
cases leading to a larger number of unique profiles being generated.

Also switched the name of the kernel module list data profile to
better conform to the proposed suseconnect changes.

Update the generate-hwinfo Makefile rule to generate '_' (underscore)
prefixed directory names so that they are ignored by the Golang ./...
recursive processing directive. Avoids long delays in Makefile rules
such as build and vet from trying to resursively analyize 100000+
entry client data store.

Added a Makefile rule, num-rmt-clients, to report the current number
of registered RMT clients.

Switch to using the SUSE/connect-ng mainstream rather than the next
branch.

Update README.md to reflect recent changes.